refactor: 폴더 `shared` -> `base` 변경, `CartIcon` 상속
