+++ src/components/Header/HeaderCartIcon.jsx
@@ -1,0 +1,7 @@
+import { StyledHeaderCartIcon } from 'components/header/style';
+
+const HeaderCartIcon = ({ ...rest }) => {
+  return <StyledHeaderCartIcon {...rest} />;
+};
+
+export default HeaderCartIcon;
