+++ src/components/base/button/Button.jsx
@@ -1,4 +1,4 @@
-import StyledButton from 'components/shared/button/style';
+import StyledButton from 'components/base/button/style';
 
 const Button = ({ children, ...rest }) => {
   return <StyledButton {...rest}>{children}</StyledButton>;
