+++ src/components/base/cartIcon/style.js
@@ -1,4 +1,4 @@
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import { ReactComponent as ShoppingCart } from 'assets/shopping_cart_icon.svg';
 
 const StyledCartIcon = styled(ShoppingCart)`
@@ -8,16 +8,6 @@ const StyledCartIcon = styled(ShoppingCart)`
   width: 30px;
   height: 26px;
   cursor: pointer;
-
-  ${({ theme }) =>
-    theme === 'header' &&
-    css`
-      path {
-        fill: white;
-      }
-      width: 49px;
-      height: 39px;
-    `}
 `;
 
 export default StyledCartIcon;
