+++ src/components/base/modal/Modal.jsx
@@ -1,4 +1,4 @@
-import StyledModal from 'components/shared/modal/style';
+import StyledModal from 'components/base/modal/style';
 
 const Modal = ({ children, ...rest }) => {
   return <StyledModal {...rest}>{children}</StyledModal>;
