+++ src/components/base/text/style.js
